# swagger/swagger.yaml
...
info:
  contact: {}
  description: 请输入以 `Bearer ` 开头的 Token，例：Bearer xxxxxxxx。未列出接口请参考前端调用方法。Token在个人中心-API密钥菜单下申请。
  title: k8m API
  version: "1.0"
paths:
...
  /k8s/cluster/{cluster}/helm/release/{release}/repo/{repo}/chart/{chart}/version/{version}/install:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: Release名称
        in: path
        name: release
        required: true
        type: string
      - description: 仓库名称
        in: path
        name: repo
        required: true
        type: string
      - description: Chart名称
        in: path
        name: chart
        required: true
        type: string
      - description: 版本号
        in: path
        name: version
        required: true
        type: string
      - description: 安装参数
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 安装Helm Release
  /k8s/cluster/{cluster}/helm/release/batch/uninstall:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 批量卸载参数
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 批量卸载Helm Release
  /k8s/cluster/{cluster}/helm/release/list:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Release列表
  /k8s/cluster/{cluster}/helm/release/ns/{ns}/name/{name}/history/list:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: Release名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取Release的历史版本
  /k8s/cluster/{cluster}/helm/release/ns/{ns}/name/{name}/revision/{revision}/notes:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: Release名称
        in: path
        name: name
        required: true
        type: string
      - description: 版本号
        in: path
        name: revision
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取ReleaseNote
  /k8s/cluster/{cluster}/helm/release/ns/{ns}/name/{name}/revision/{revision}/values:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: Release名称
        in: path
        name: name
        required: true
        type: string
      - description: 版本号
        in: path
        name: revision
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取安装yaml
  /k8s/cluster/{cluster}/helm/release/ns/{ns}/name/{name}/uninstall:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 命名空间
        in: path
        name: ns
        required: true
        type: string
      - description: Release名称
        in: path
        name: name
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 卸载Helm Release
  /k8s/cluster/{cluster}/helm/release/upgrade:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 升级参数
        in: body
        name: body
        required: true
        schema:
          type: object
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 升级Helm Release
...
  /k8s/cluster/{cluster}/k8s_gpt/cluster/{user_cluster}/result:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 用户集群标识
        in: path
        name: user_cluster
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取集群K8sGPT分析结果
  /k8s/cluster/{cluster}/k8s_gpt/cluster/{user_cluster}/run:
    post:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 用户集群标识
        in: path
        name: user_cluster
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 对整个集群运行K8sGPT分析
  /k8s/cluster/{cluster}/k8s_gpt/kind/{kind}/run:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      - description: 资源类型
        in: path
        name: kind
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 对指定资源类型运行K8sGPT分析
  /k8s/cluster/{cluster}/k8s_gpt/var:
    get:
      parameters:
      - description: 集群名称
        in: query
        name: cluster
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: 获取K8s资源字段信息
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"